refactor : Change component directory structure
