+++ src/components/ui/atoms/Button/index.tsx
@@ -1,0 +1,3 @@
+/* Button 컴포넌트와 관련 타입들을 외부로 내보내는 진입점 파일 */
+export { Button } from './Button';
+export type { ButtonProps } from './Button';
