+++ src/components/ui/atoms/Icon/index.tsx
@@ -1,0 +1,2 @@
+export { Icon } from './Icon';
+export type { IconProps } from './Icon';
