+++ src/components/ui/atoms/Label/index.tsx
@@ -1,0 +1,2 @@
+export { Label } from './Label';
+export type { LabelProps } from './Label';
