+++ src/components/ui/atoms/Tag/index.tsx
@@ -1,0 +1,2 @@
+export { Tag } from './Tag';
+export type { TagProps } from './Tag';
