+++ src/components/ui/atoms/Typography/index.tsx
@@ -1,0 +1,6 @@
+export { Heading } from './Typography';
+export { Text } from './Typography';
+export { DateText } from './Typography';
+export { Caption } from './Typography';
+export { Italic } from './Typography';
+export type { TypographyProps } from './Typography';
