+++ src/components/ui/molecules/BlogPostCard/index.tsx
@@ -1,0 +1,2 @@
+export { BlogPostCard } from './BlogPostCard';
+export type { BlogPostCardProps } from './BlogPostCard';
