+++ src/components/ui/molecules/PaginationBar/index.tsx
@@ -1,0 +1,2 @@
+export { PaginationBar } from './PaginationBar';
+export type { PaginationBarProps } from './PaginationBar';
